refactor: convert `rullate` from Long to int in CalciobalillaController

+++ BitPub/BitPub-Cloud/src/main/java/com/bitpub/controllers/CalciobalillaController.java
@@ -64,7 +64,7 @@ public class CalciobalillaController {
      */
     @GetMapping("/stats")
     public ResponseEntity<CalciobalillaStats> getGlobalStats() {
-        Long rullate = repository.countTotalRullate() != null ? repository.countTotalRullate() : 0;
+        int rullate = repository.countTotalRullate() != null ? repository.countTotalRullate() : 0;
         int vRossi = repository.countVittorieRossi();
         int vBlu = repository.countVittorieBlu();
